Add GSSAPI/Kerberos authentication via Negotiate
When the service name is set via the servicename config option and
pykerberos is installed allow authentication via the negotiate header.

Since this is not using basic auth and it's on top of all other
authenciation schemes it's not implemented as an acl module. This will
also allow us to make the whole negotiate auth be connection based in
the future.

This has been succesfully tested with iceowl.